refactor user interface for command line
