+++ lib/gateway/in_memory_game.rb
@@ -1,0 +1,11 @@
+class Gateway::InMemoryGame
+  attr_accessor :saved_game
+
+  def retrieve
+    @saved_game.dup
+  end
+
+  def save(game)
+    @saved_game = game
+  end
+end
